requirements.txt
mediapipe
opencv-python
numpy
pyautogui
pillow